Move util functions to imod5_converter utilities
